// src/types/MedicalScheme.ts
export interface MedicalScheme {
  MedicalSchemeId: number;
  TotalMonthlyCostMainMember: number;
  MedicalSavingsAccountMainMember: number;
  TotalMonthlyCostAdultDependant: number;
  MedicalSavingsAccountAdultDependant: number;
  TotalMonthlyCostChildDependant: number;
  MedicalSavingsAccountChildDependant: number;
}